verandering van naam RoomEditor naam RoomTemplate + view en viewmodel
