+++ Designer/View/RoomEditorView.xaml.cs
@@ -1,24 +1,26 @@
-﻿using Designer.Other;
-using Designer.ViewModel;
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
 
-namespace Designer.View {
+namespace Designer.View
+{
     /// <summary>
     /// Interaction logic for RoomEditorView.xaml
     /// </summary>
-    public partial class RoomEditorView : Page {
-
-        public BasicCommand GotoRoomTemplate { get; set; }
-        public RoomEditorView() {
-            //pagina initializen
-
+    public partial class RoomEditorView : Page
+    {
+        public RoomEditorView()
+        {
             InitializeComponent();
-
-            GotoRoomTemplate = new PageCommand(() => new RoomOverview());
-
         }
-
-
-
     }
 }
