+++ DesignerTest/RoomEditorTests.cs
@@ -14,7 +14,7 @@ namespace DesignerTest {
         [TestCase("kamernaam_test4", 340, 69)]
         public void SaveRoomWithoutTemplate_SameAsReturn(string name, int width, int length) {
             // deze methode slaat de kamer op.
-            Room room = RoomEditorViewModel.SaveRoom(name, width, length);
+            Room room = RoomTemplateViewModel.SaveRoom(name, width, length);
             
             Assert.NotNull(room);
             Assert.AreEqual(room.Name, name);
@@ -36,7 +36,7 @@ namespace DesignerTest {
         public void SaveRoomWithTemplate_SameAsReturn(string name, int width, int length, int template)
         {
             // deze methode slaat de kamer op.
-            Room room = RoomEditorViewModel.SaveRoom(name, width, length, template);
+            Room room = RoomTemplateViewModel.SaveRoom(name, width, length, template);
 
             Assert.NotNull(room);
             Assert.AreEqual(room.Name, name);
